Refactor withdraw tests more
