+++ stake-pool/program/tests/withdraw_edge_cases.rs
@@ -10,101 +10,9 @@ use {
         borsh::try_from_slice_unchecked, instruction::InstructionError, pubkey::Pubkey, stake,
     },
     solana_program_test::*,
-    solana_sdk::{
-        signature::{Keypair, Signer},
-        transaction::TransactionError,
-    },
-    spl_stake_pool::{
-        error::StakePoolError, instruction, minimum_stake_lamports, state, MINIMUM_RESERVE_LAMPORTS,
-    },
+    solana_sdk::{signature::Signer, transaction::TransactionError},
+    spl_stake_pool::{error::StakePoolError, instruction, state, MINIMUM_RESERVE_LAMPORTS},
 };
-
-async fn setup(
-    token_program_id: Pubkey,
-) -> (
-    ProgramTestContext,
-    StakePoolAccounts,
-    ValidatorStakeAccount,
-    DepositStakeAccount,
-    Keypair,
-    Keypair,
-    u64,
-) {
-    let mut context = program_test().start_with_context().await;
-    let stake_pool_accounts = StakePoolAccounts::new_with_token_program(token_program_id);
-    stake_pool_accounts
-        .initialize_stake_pool(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            MINIMUM_RESERVE_LAMPORTS,
-        )
-        .await
-        .unwrap();
-
-    let validator_stake_account = simple_add_validator_to_pool(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        None,
-    )
-    .await;
-
-    let current_minimum_delegation = stake_pool_get_minimum_delegation(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-    )
-    .await;
-
-    let deposit_info = simple_deposit_stake(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts,
-        &validator_stake_account,
-        current_minimum_delegation * 3,
-    )
-    .await
-    .unwrap();
-
-    let tokens_to_withdraw = deposit_info.pool_tokens;
-
-    // Delegate tokens for withdrawing
-    let user_transfer_authority = Keypair::new();
-    delegate_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &deposit_info.pool_account.pubkey(),
-        &deposit_info.authority,
-        &user_transfer_authority.pubkey(),
-        tokens_to_withdraw,
-    )
-    .await;
-
-    // Create stake account to withdraw to
-    let user_stake_recipient = Keypair::new();
-    create_blank_stake_account(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &user_stake_recipient,
-    )
-    .await;
-
-    (
-        context,
-        stake_pool_accounts,
-        validator_stake_account,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_withdraw,
-    )
-}
 
 #[tokio::test]
 async fn fail_remove_validator() {
@@ -116,7 +24,7 @@ async fn fail_remove_validator() {
         user_transfer_authority,
         user_stake_recipient,
         _,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     // decrease a little stake, not all
     let error = stake_pool_accounts
@@ -189,7 +97,7 @@ async fn success_remove_validator() {
         user_transfer_authority,
         user_stake_recipient,
         _,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let rent = context.banks_client.get_rent().await.unwrap();
     let stake_rent = rent.minimum_balance(std::mem::size_of::<stake::state::StakeState>());
@@ -292,7 +200,7 @@ async fn fail_with_reserve() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     // decrease a little stake, not all
     let error = stake_pool_accounts
@@ -362,7 +270,7 @@ async fn success_with_reserve() {
         user_transfer_authority,
         user_stake_recipient,
         _,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let rent = context.banks_client.get_rent().await.unwrap();
     let stake_rent = rent.minimum_balance(std::mem::size_of::<stake::state::StakeState>());
@@ -483,7 +391,7 @@ async fn success_with_empty_preferred_withdraw() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let preferred_validator = simple_add_validator_to_pool(
         &mut context.banks_client,
@@ -532,7 +440,7 @@ async fn success_and_fail_with_preferred_withdraw() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_burn,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     let preferred_validator = simple_add_validator_to_pool(
         &mut context.banks_client,
@@ -615,7 +523,7 @@ async fn fail_withdraw_from_transient() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     // add a preferred withdraw validator, keep it empty, to be sure that this works
     let preferred_validator = simple_add_validator_to_pool(
@@ -690,7 +598,7 @@ async fn success_withdraw_from_transient() {
         user_transfer_authority,
         user_stake_recipient,
         tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
+    ) = setup_for_withdraw(spl_token::id()).await;
 
     // add a preferred withdraw validator, keep it empty, to be sure that this works
     let preferred_validator = simple_add_validator_to_pool(
@@ -746,75 +654,4 @@ async fn success_withdraw_from_transient() {
         )
         .await;
     assert!(error.is_none());
-}
-
-#[tokio::test]
-async fn success_withdraw_all_fee_tokens() {
-    let (
-        mut context,
-        stake_pool_accounts,
-        validator_stake_account,
-        deposit_info,
-        user_transfer_authority,
-        user_stake_recipient,
-        tokens_to_withdraw,
-    ) = setup(spl_token::id()).await;
-
-    // move tokens to fee account
-    transfer_spl_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &deposit_info.pool_account.pubkey(),
-        &stake_pool_accounts.pool_mint.pubkey(),
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-        &user_transfer_authority,
-        tokens_to_withdraw / 2,
-        stake_pool_accounts.pool_decimals,
-    )
-    .await;
-
-    let fee_tokens = get_token_balance(
-        &mut context.banks_client,
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-    )
-    .await;
-
-    let user_transfer_authority = Keypair::new();
-    delegate_tokens(
-        &mut context.banks_client,
-        &context.payer,
-        &context.last_blockhash,
-        &stake_pool_accounts.token_program_id,
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-        &stake_pool_accounts.manager,
-        &user_transfer_authority.pubkey(),
-        fee_tokens,
-    )
-    .await;
-
-    let new_authority = Pubkey::new_unique();
-    let error = stake_pool_accounts
-        .withdraw_stake(
-            &mut context.banks_client,
-            &context.payer,
-            &context.last_blockhash,
-            &user_stake_recipient.pubkey(),
-            &user_transfer_authority,
-            &stake_pool_accounts.pool_fee_account.pubkey(),
-            &validator_stake_account.stake_account,
-            &new_authority,
-            fee_tokens,
-        )
-        .await;
-    assert!(error.is_none());
-
-    // Check balance is 0
-    let fee_tokens = get_token_balance(
-        &mut context.banks_client,
-        &stake_pool_accounts.pool_fee_account.pubkey(),
-    )
-    .await;
-    assert_eq!(fee_tokens, 0);
 }
